ML-pipeline/requirements.txt
requests==2.31.0
fastapi==0.109.2
firebase_admin
python-dotenv==1.0.0
sentry-sdk
pydub==0.25.1
uvicorn
langchain==0.0.335
pydantic==1.10.9
moviepy==1.0.3
googletrans==4.0.0-rc1
pyannote.audio
speechbrain
audiostretchy
openai-whisper
git+https://github.com/openai/whisper.git
TTS
torchaudio
ffmpeg
